homework 2.2 mvvm

+++ app/src/main/java/com/example/nmedia/Post.kt
@@ -8,8 +8,8 @@ data class Post (
     val content: String,
     val published: String,
     val likedByMe: Boolean = false,
-    val likes: Int = 0,
-    val reposts: Int = 0,
+    val likes: Long = 0,
+    val reposts: Long = 0,
     //var countReposts: Int = 0,
    // var countLikes: Int = 1
 )
